# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION for Geotager",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "prosenjak.urban@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get all locations",
                "description": "Fetch all locations with pagination",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 40
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Locations fetched successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create location",
                "description": "Create a new location with image upload",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude",
                                    "user_id"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "Location image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "latitude": {
                                        "description": "Location latitude",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "description": "Location longitude",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "user_id": {
                                        "description": "User ID associated with location",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location created successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get a single location",
                "description": "Fetch a single location by its ID",
                "operationId": "findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update location",
                "description": "Update an existing location with optional image upload",
                "operationId": "updateLocation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Location image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "latitude": {
                                        "description": "Location latitude",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "description": "Location longitude",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "user_id": {
                                        "description": "User ID associated with location",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete location",
                "description": "Delete an existing location and its associated image",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/location/random": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get a random location",
                "description": "Fetch a single random location",
                "operationId": "random",
                "responses": {
                    "200": {
                        "description": "Random location fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No locations available"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/locations/{user_id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get locations by User ID",
                "description": "Fetch all locations associated with a specific user ID with pagination",
                "operationId": "getLocationsByUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user whose locations are to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Locations fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Location"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Location": {
                "title": "Location",
                "required": [
                    "id",
                    "latitude",
                    "longitude",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "Location ID",
                        "type": "integer"
                    },
                    "image": {
                        "description": "Location image",
                        "type": "string",
                        "format": "binary"
                    },
                    "latitude": {
                        "description": "Latitude",
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "description": "Longitude",
                        "type": "number",
                        "format": "float"
                    },
                    "user_id": {
                        "description": "User ID associated with the location",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}